+++ pkg/providers/cloudlets/create_policy.go
@@ -12,6 +12,7 @@ import (
 	"text/template"
 
 	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v7/pkg/cloudlets"
+	v3 "github.com/akamai/AkamaiOPEN-edgegrid-golang/v7/pkg/cloudlets/v3"
 	"github.com/akamai/cli-terraform/pkg/edgegrid"
 	"github.com/akamai/cli-terraform/pkg/templates"
 	"github.com/akamai/cli-terraform/pkg/tools"
@@ -29,11 +30,12 @@ type (
 		Description             string
 		GroupID                 int64
 		MatchRuleFormat         cloudlets.MatchRuleFormat
-		MatchRules              cloudlets.MatchRules
-		PolicyActivations       map[string]TFPolicyActivationData
+		MatchRules              any
+		PolicyActivations       TFPolicyActivationsData
 		LoadBalancers           []LoadBalancerVersion
 		LoadBalancerActivations []cloudlets.LoadBalancerActivation
 		Section                 string
+		IsV3                    bool
 	}
 	// LoadBalancerVersion ...
 	LoadBalancerVersion struct {
@@ -41,27 +43,43 @@ type (
 		OriginDescription string
 	}
 
+	// TFPolicyActivationsData represents data used in policy all activation resource templates
+	TFPolicyActivationsData struct {
+		Staging    *TFPolicyActivationData
+		Production *TFPolicyActivationData
+		IsV3       bool
+	}
+
 	// TFPolicyActivationData represents data used in policy activation resource templates
 	TFPolicyActivationData struct {
 		PolicyID   int64
 		Version    int64
 		Properties []string
+		IsV3       bool
+	}
+
+	activationStrategy interface {
+		initializeWithPolicy(ctx context.Context, policyName string) error
+		validatePolicy() error
+		populateWithLatestPolicyVersion(ctx context.Context) error
+		getTFPolicyData(ctx context.Context, section string) (*TFPolicyData, error)
+	}
+
+	v2ActivationStrategy struct {
+		client        cloudlets.Cloudlets
+		policy        *cloudlets.Policy
+		policyVersion *cloudlets.PolicyVersion
+	}
+
+	v3ActivationStrategy struct {
+		client        v3.Cloudlets
+		policy        *v3.Policy
+		policyVersion *v3.PolicyVersion
 	}
 )
 
 //go:embed templates/*
 var templateFiles embed.FS
-
-var supportedCloudlets = map[string]struct{}{
-	"ALB": {},
-	"AP":  {},
-	"AS":  {},
-	"CD":  {},
-	"ER":  {},
-	"FR":  {},
-	"IG":  {},
-	"VP":  {},
-}
 
 var (
 	// ErrFetchingPolicy is returned when fetching policy fails
@@ -70,13 +88,17 @@ var (
 	ErrFetchingVersion = errors.New("unable to fetch latest policy version")
 	// ErrCloudletTypeNotSupported is returned when a provided cloudlet type is not yet supported
 	ErrCloudletTypeNotSupported = errors.New("cloudlet type not supported")
+
+	errPolicyNotFound   = errors.New("policy does not exist")
+	errVersionsNotFound = errors.New("no policy versions found for given policy")
 )
 
 // CmdCreatePolicy is an entrypoint to create-policy command
 func CmdCreatePolicy(c *cli.Context) error {
 	ctx := c.Context
 	sess := edgegrid.GetSession(c.Context)
-	client := cloudlets.Client(sess)
+	clientV2 := cloudlets.Client(sess)
+	clientV3 := v3.Client(sess)
 
 	// tfWorkPath is a target directory for generated terraform resources
 	var tfWorkPath = "./"
@@ -112,81 +134,102 @@ func CmdCreatePolicy(c *cli.Context) error {
 
 	policyName := c.Args().First()
 	section := edgegrid.GetEdgercSection(c)
-	if err = createPolicy(ctx, policyName, section, client, processor); err != nil {
+	if err = createPolicy(ctx, policyName, section, clientV2, clientV3, processor); err != nil {
 		return cli.Exit(color.RedString(fmt.Sprintf("Error exporting policy HCL: %s", err)), 1)
 	}
 	return nil
 }
 
-func createPolicy(ctx context.Context, policyName, section string, client cloudlets.Cloudlets, templateProcessor templates.TemplateProcessor) error {
+func createPolicy(ctx context.Context, policyName, section string, clientV2 cloudlets.Cloudlets, clientV3 v3.Cloudlets, templateProcessor templates.TemplateProcessor) error {
 	term := terminal.Get(ctx)
 
 	fmt.Println("Configuring Policy")
 	term.Spinner().Start("Fetching policy " + policyName)
 
-	policy, err := findPolicyByName(ctx, policyName, client)
+	strategy, err := initializeStrategyForPolicy(ctx, policyName, clientV2, clientV3)
 	if err != nil {
 		term.Spinner().Fail()
 		return fmt.Errorf("%w: %s", ErrFetchingPolicy, err)
 	}
-	if _, ok := supportedCloudlets[policy.CloudletCode]; !ok {
+	if err = strategy.validatePolicy(); err != nil {
 		term.Spinner().Fail()
-		return fmt.Errorf("%w: %s", ErrCloudletTypeNotSupported, policy.CloudletCode)
+		return err
 	}
 
-	tfPolicyData := TFPolicyData{
-		Section:      section,
-		Name:         policy.Name,
-		CloudletCode: policy.CloudletCode,
-		GroupID:      policy.GroupID,
-	}
-
-	policyVersion, err := getLatestPolicyVersion(ctx, policy.PolicyID, client)
+	err = strategy.populateWithLatestPolicyVersion(ctx)
 	if err != nil {
 		term.Spinner().Fail()
 		return fmt.Errorf("%w: %s", ErrFetchingVersion, err)
 	}
-	tfPolicyData.Description = policyVersion.Description
-	tfPolicyData.MatchRuleFormat = policyVersion.MatchRuleFormat
-	tfPolicyData.MatchRules = policyVersion.MatchRules
 
-	tfPolicyData.PolicyActivations = make(map[string]TFPolicyActivationData)
-	if activationStaging := getActiveVersionAndProperties(policy, cloudlets.PolicyActivationNetworkStaging); activationStaging != nil {
-		tfPolicyData.PolicyActivations["staging"] = *activationStaging
-	}
-	if activationProd := getActiveVersionAndProperties(policy, cloudlets.PolicyActivationNetworkProduction); activationProd != nil {
-		tfPolicyData.PolicyActivations["prod"] = *activationProd
-	}
-
-	if tfPolicyData.CloudletCode == "ALB" {
-		originIDs, err := getOriginIDs(policyVersion.MatchRules)
-		if err != nil {
-			term.Spinner().Fail()
-			return fmt.Errorf("%w: %s", ErrFetchingVersion, err)
-		}
-		tfPolicyData.LoadBalancers, err = getLoadBalancers(ctx, client, originIDs)
-		if err != nil {
-			term.Spinner().Fail()
-			return fmt.Errorf("%w: %s", ErrFetchingVersion, err)
-		}
-		tfPolicyData.LoadBalancerActivations, err = getLoadBalancerActivations(ctx, client, originIDs)
-		if err != nil {
-			term.Spinner().Fail()
-			return fmt.Errorf("%w: %s", ErrFetchingVersion, err)
-		}
-
+	tfPolicyData, err := strategy.getTFPolicyData(ctx, section)
+	if err != nil {
+		term.Spinner().Fail()
+		return err
 	}
 
 	term.Spinner().OK()
 	term.Spinner().Start("Saving TF configurations ")
-	if err := templateProcessor.ProcessTemplates(tfPolicyData); err != nil {
+	if err := templateProcessor.ProcessTemplates(*tfPolicyData); err != nil {
 		term.Spinner().Fail()
 		return err
 	}
 	term.Spinner().OK()
-	fmt.Printf("Terraform configuration for policy '%s' was saved successfully\n", policy.Name)
+	fmt.Printf("Terraform configuration for policy '%s' was saved successfully\n", policyName)
 
 	return nil
+}
+
+func (strategy *v2ActivationStrategy) getTFPolicyData(ctx context.Context, section string) (*TFPolicyData, error) {
+	tfPolicyData := TFPolicyData{
+		Section:           section,
+		Name:              strategy.policy.Name,
+		CloudletCode:      strategy.policy.CloudletCode,
+		GroupID:           strategy.policy.GroupID,
+		Description:       strategy.policyVersion.Description,
+		MatchRuleFormat:   strategy.policyVersion.MatchRuleFormat,
+		MatchRules:        strategy.policyVersion.MatchRules,
+		PolicyActivations: TFPolicyActivationsData{IsV3: false},
+	}
+
+	if activationStaging := getActiveVersionAndProperties(strategy.policy, cloudlets.PolicyActivationNetworkStaging); activationStaging != nil {
+		tfPolicyData.PolicyActivations.Staging = activationStaging
+	}
+	if activationProd := getActiveVersionAndProperties(strategy.policy, cloudlets.PolicyActivationNetworkProduction); activationProd != nil {
+		tfPolicyData.PolicyActivations.Production = activationProd
+	}
+
+	if tfPolicyData.CloudletCode == "ALB" {
+		originIDs, err := getOriginIDs(strategy.policyVersion.MatchRules)
+		if err != nil {
+
+			return nil, fmt.Errorf("%w: %s", ErrFetchingVersion, err)
+		}
+		tfPolicyData.LoadBalancers, err = getLoadBalancers(ctx, strategy.client, originIDs)
+		if err != nil {
+			return nil, fmt.Errorf("%w: %s", ErrFetchingVersion, err)
+		}
+		tfPolicyData.LoadBalancerActivations, err = getLoadBalancerActivations(ctx, strategy.client, originIDs)
+		if err != nil {
+			return nil, fmt.Errorf("%w: %s", ErrFetchingVersion, err)
+		}
+
+	}
+	return &tfPolicyData, nil
+}
+
+func initializeStrategyForPolicy(ctx context.Context, policyName string, clientV2 cloudlets.Cloudlets, clientV3 v3.Cloudlets) (activationStrategy, error) {
+	v2Strategy := v2ActivationStrategy{client: clientV2}
+	errV2 := v2Strategy.initializeWithPolicy(ctx, policyName)
+	if errV2 == nil {
+		return &v2Strategy, nil
+	}
+	v3Strategy := v3ActivationStrategy{client: clientV3}
+	errV3 := v3Strategy.initializeWithPolicy(ctx, policyName)
+	if errV3 == nil {
+		return &v3Strategy, nil
+	}
+	return nil, fmt.Errorf("could not find policy %s: neither as V2 (%s) nor as V3 (%s)", policyName, errV2, errV3)
 }
 
 func getLoadBalancerActivations(ctx context.Context, client cloudlets.Cloudlets, originIDs []string) ([]cloudlets.LoadBalancerActivation, error) {
@@ -292,21 +335,20 @@ func getApplicationLoadBalancerActivation(ctx context.Context, client cloudlets.
 	return nil, nil
 }
 
-func findPolicyByName(ctx context.Context, name string, client cloudlets.Cloudlets) (*cloudlets.Policy, error) {
+func (strategy *v2ActivationStrategy) initializeWithPolicy(ctx context.Context, name string) error {
 	pageSize, offset := 1000, 0
-	var policy *cloudlets.Policy
 	for {
-		policies, err := client.ListPolicies(ctx, cloudlets.ListPoliciesRequest{
+		policies, err := strategy.client.ListPolicies(ctx, cloudlets.ListPoliciesRequest{
 			Offset:   offset,
 			PageSize: &pageSize,
 		})
 		if err != nil {
-			return nil, err
+			return err
 		}
 		for _, p := range policies {
 			if p.Name == name {
-				policy = &p
-				return policy, nil
+				strategy.policy = &p
+				return nil
 			}
 		}
 		if len(policies) < pageSize {
@@ -314,25 +356,26 @@ func findPolicyByName(ctx context.Context, name string, client cloudlets.Cloudle
 		}
 		offset += pageSize
 	}
-	return nil, fmt.Errorf("policy '%s' does not exist", name)
+	return errPolicyNotFound
 }
 
-func getLatestPolicyVersion(ctx context.Context, policyID int64, client cloudlets.Cloudlets) (*cloudlets.PolicyVersion, error) {
+func (strategy *v2ActivationStrategy) populateWithLatestPolicyVersion(ctx context.Context) error {
 	var version int64
+	policyID := strategy.policy.PolicyID
 	pageSize, offset := 1000, 0
 	for {
-		versions, err := client.ListPolicyVersions(ctx, cloudlets.ListPolicyVersionsRequest{
+		versions, err := strategy.client.ListPolicyVersions(ctx, cloudlets.ListPolicyVersionsRequest{
 			PolicyID:     policyID,
 			IncludeRules: false,
 			PageSize:     &pageSize,
 			Offset:       offset,
 		})
 		if err != nil {
-			return nil, err
+			return err
 		}
 
 		if len(versions) == 0 {
-			return nil, fmt.Errorf("no policy versions found for given policy")
+			return errVersionsNotFound
 		}
 		for _, v := range versions {
 			if v.Version > version {
@@ -344,14 +387,15 @@ func getLatestPolicyVersion(ctx context.Context, policyID int64, client cloudlet
 		}
 		offset += pageSize
 	}
-	policyVersion, err := client.GetPolicyVersion(ctx, cloudlets.GetPolicyVersionRequest{
+	policyVersion, err := strategy.client.GetPolicyVersion(ctx, cloudlets.GetPolicyVersionRequest{
 		PolicyID: policyID,
 		Version:  version,
 	})
 	if err != nil {
-		return nil, err
+		return err
 	}
-	return policyVersion, nil
+	strategy.policyVersion = policyVersion
+	return nil
 }
 
 func getActiveVersionAndProperties(policy *cloudlets.Policy, network cloudlets.PolicyActivationNetwork) *TFPolicyActivationData {
@@ -371,5 +415,129 @@ func getActiveVersionAndProperties(policy *cloudlets.Policy, network cloudlets.P
 		PolicyID:   policy.PolicyID,
 		Version:    version,
 		Properties: associatedProperties,
+	}
+}
+
+func (strategy *v2ActivationStrategy) validatePolicy() error {
+	var supportedCloudlets = map[string]struct{}{
+		"ALB": {},
+		"AP":  {},
+		"AS":  {},
+		"CD":  {},
+		"ER":  {},
+		"FR":  {},
+		"IG":  {},
+		"VP":  {},
+	}
+
+	_, ok := supportedCloudlets[strategy.policy.CloudletCode]
+	if !ok {
+		return fmt.Errorf("%w: %s", ErrCloudletTypeNotSupported, strategy.policy.CloudletCode)
+	}
+	return nil
+}
+
+func (strategy *v3ActivationStrategy) initializeWithPolicy(ctx context.Context, policyName string) error {
+	page, size := 0, 1000
+	for {
+		policies, err := strategy.client.ListPolicies(ctx, v3.ListPoliciesRequest{Page: page, Size: size})
+		if err != nil {
+			return err
+		}
+		for _, policy := range policies.Content {
+			if policy.Name == policyName {
+				strategy.policy = &policy
+				return nil
+			}
+		}
+		if len(policies.Content) < size {
+			break
+		}
+		page++
+	}
+	return errPolicyNotFound
+}
+
+func (strategy *v3ActivationStrategy) validatePolicy() error {
+	var supportedCloudlets = map[v3.CloudletType]struct{}{
+		v3.CloudletTypeAP: {},
+		v3.CloudletTypeAS: {},
+		v3.CloudletTypeCD: {},
+		v3.CloudletTypeER: {},
+		v3.CloudletTypeFR: {},
+		v3.CloudletTypeIG: {},
+	}
+
+	_, ok := supportedCloudlets[strategy.policy.CloudletType]
+	if !ok {
+		return fmt.Errorf("%w: %s", ErrCloudletTypeNotSupported, strategy.policy.CloudletType)
+	}
+	return nil
+}
+
+func (strategy *v3ActivationStrategy) populateWithLatestPolicyVersion(ctx context.Context) error {
+	policyID := strategy.policy.ID
+
+	versions, err := strategy.client.ListPolicyVersions(ctx, v3.ListPolicyVersionsRequest{PolicyID: policyID, Page: 0, Size: 10})
+	if err != nil {
+		return err
+	}
+	if len(versions.PolicyVersions) == 0 {
+		return errVersionsNotFound
+	}
+
+	policyVersion, err := strategy.client.GetPolicyVersion(ctx, v3.GetPolicyVersionRequest{
+		PolicyID: policyID,
+		Version:  versions.PolicyVersions[0].Version,
+	})
+	if err != nil {
+		return err
+	}
+	strategy.policyVersion = policyVersion
+	return nil
+}
+
+func (strategy *v3ActivationStrategy) getTFPolicyData(_ context.Context, section string) (*TFPolicyData, error) {
+	tfPolicyData := TFPolicyData{
+		Section:           section,
+		Name:              strategy.policy.Name,
+		CloudletCode:      string(strategy.policy.CloudletType),
+		GroupID:           strategy.policy.GroupID,
+		IsV3:              true,
+		PolicyActivations: TFPolicyActivationsData{IsV3: true},
+	}
+
+	if strategy.policyVersion.Description != nil {
+		tfPolicyData.Description = *strategy.policyVersion.Description
+	}
+	if strategy.policyVersion.MatchRules != nil {
+		tfPolicyData.MatchRules = strategy.policyVersion.MatchRules
+	}
+
+	if activationStaging := strategy.getActivationDataForV3(v3.StagingNetwork); activationStaging != nil {
+		tfPolicyData.PolicyActivations.Staging = activationStaging
+	}
+	if activationProd := strategy.getActivationDataForV3(v3.ProductionNetwork); activationProd != nil {
+		tfPolicyData.PolicyActivations.Production = activationProd
+	}
+
+	return &tfPolicyData, nil
+}
+
+func (strategy *v3ActivationStrategy) getActivationDataForV3(network v3.Network) *TFPolicyActivationData {
+	var activationToCheck *v3.PolicyActivation
+	switch network {
+	case v3.StagingNetwork:
+		activationToCheck = strategy.policy.CurrentActivations.Staging.Effective
+	case v3.ProductionNetwork:
+		activationToCheck = strategy.policy.CurrentActivations.Production.Effective
+	}
+	if activationToCheck == nil || activationToCheck.Operation == v3.OperationDeactivation {
+		return nil
+	}
+	return &TFPolicyActivationData{
+		PolicyID: activationToCheck.PolicyID,
+		Version:  activationToCheck.PolicyVersion,
+		IsV3:     true,
 	}
 }