+++ pkg/providers/cloudlets/testdata/v3/v3_with_both_activations/import.sh
@@ -1,0 +1,3 @@
+terraform init
+terraform import akamai_cloudlets_policy.policy test_policy_export
+terraform import akamai_cloudlets_policy_activation.policy_activation 2:prod